+++ test/membrane/integration/endpoint_test.exs
@@ -14,8 +14,8 @@ defmodule Membrane.Core.EndpointTest do
       buffers = ['a', 'b', 'c']
 
       {:ok, pipeline} =
-        Testing.Pipeline.start_link(%Testing.Pipeline.Options{
-          elements: [
+        Testing.Pipeline.start_link(
+          children: [
             endpoint: %Testing.Endpoint{output: buffers},
             filter: TestFilter
           ],
@@ -23,7 +23,7 @@ defmodule Membrane.Core.EndpointTest do
             link(:endpoint) |> to(:filter),
             link(:filter) |> to(:endpoint)
           ]
-        })
+        )
 
       assert_data_flows_through(pipeline, buffers, :endpoint)
     end
@@ -32,8 +32,8 @@ defmodule Membrane.Core.EndpointTest do
       buffers = ['a', 'b', 'c']
 
       {:ok, pipeline} =
-        Testing.Pipeline.start_link(%Testing.Pipeline.Options{
-          elements: [
+        Testing.Pipeline.start_link(
+          children: [
             endpoint: %Testing.Endpoint{output: buffers},
             filter1: TestFilter,
             filter2: TestFilter,
@@ -45,7 +45,7 @@ defmodule Membrane.Core.EndpointTest do
             link(:filter2) |> to(:filter3),
             link(:filter3) |> to(:endpoint)
           ]
-        })
+        )
 
       assert_data_flows_through(pipeline, buffers, :endpoint)
     end